added second component
